fix(datastore): Fix enum member field renaming
When generating serialisation implementations, serde only applies the rename_all
attribute to the members of an enum itself - but not to their struct fields.

This changes the code to add explicit renaming attributes on the struct fields
where necessary. Hopefully this will become unnecessary in the future as I have
filed an issue with Serde upstream.